requirements.txt
PyYAML
pyperclip
pyotp
